requirements.txt
# =====================================================
# AI Voice Platform v2 - Python Dependencies
# =====================================================

# FastAPI & Web Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-multipart==0.0.6

# Database & ORM
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
redis==5.0.1

# Authentication & Security
pyjwt==2.2.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Twilio
twilio==8.11.0

# Deepgram STT
deepgram-sdk==3.2.0
pydub==0.25.1

# ElevenLabs TTS
elevenlabs==1.0.0

# OpenAI
openai==1.10.0

# Microsoft Graph
msal==1.24.1
requests==2.31.0

# Data Processing
pydantic==2.5.3
pydantic-settings==2.1.0
python-dateutil==2.8.2
pytz==2024.1
pyyaml==6.0.1
psutil==5.9.8

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Task Queue (for background jobs)
celery==5.3.6
kombu==5.3.5

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3

# Testing (dev)
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Code Quality
ruff==0.1.14
mypy==1.8.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3
